requirements.txt
PySide6==6.7.3
opencv-python==4.9.0.80
face-recognition==1.3.0
dlib==19.24.0
Pillow==10.4.0
numpy==1.26.4